hw3/hw3_R14922146_code/hw3_R14922146_code/requirements.txt
tqdm
torch
Pillow
numpy
transformers
nltk
Image
numpy
torch --index-url https://download.pytorch.org/whl/cu121
transformers==4.40.0
huggingface_hub
numpy
datasets
peft
accelerate